requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
dotenv==0.9.9
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
psycopg==3.2.6
psycopg-binary==3.2.6
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
plaid-python>=17.0.0
